requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pydantic
requests
numpy
pydantic-settings
beautifulsoup4
pytest
pinecone-client

